Improve the mystery rhythm
I still wonder if measures 16-17 should be collapsed to one measure with eighth-eighth-half rhythm (semibreve semibreve longa), but I don't really have solid evidence to justify that. In the meantime, though, if the rhythm is as notated in the original — breve breve longa — then the second breve should be altered, giving a quarter-half-dotted-half rhythm, not half-quarter-dotted-half as I initially interpreted it (I had forgotten about alteration...).
I think this sounds better, too.